# lib/api-zod/src/generated/types/billingQuote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Situation X - AI + Astrology Situation Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { BillingQuotePlanId } from "./billingQuotePlanId";

/**
 * Price breakdown of one plan. All money values are integer paise (INR).
 */
export interface BillingQuote {
  planId: BillingQuotePlanId;
  label: string;
  months: number;
  currency: string;
  monthlyPaise: number;
  grossPaise: number;
  discountPct: number;
  discountPaise: number;
  gstPct: number;
  gstPaise: number;
  totalPaise: number;
  effectivePerMonthPaise: number;
}