# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Situation X - AI + Astrology Situation Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestPlan } from "./createOrderRequestPlan";

export interface CreateOrderRequest {
  plan: CreateOrderRequestPlan;
}