# lib/api-zod/src/generated/types/createOrderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Situation X - AI + Astrology Situation Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { BillingQuote } from "./billingQuote";

export interface CreateOrderResponse {
  orderId: string;
  keyId: string;
  amountPaise: number;
  currency: string;
  plan: string;
  quote: BillingQuote;
}