# lib/api-zod/src/generated/types/subscriptionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Situation X - AI + Astrology Situation Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionStatus {
  active: boolean;
  plan: string | null;
  currentPeriodEnd: string | null;
  daysLeft: number;
  paywallEnabled: boolean;
}